SPDI-29183. Backport #14646 Fix isMounted should not invoke ufs, if not /metrics page very slowly
Fix isMounted should not invoke ufs, if not /metrics page very slowly

isMounted would not invoke ufs.

Please clarify why the changes are needed. For instance,
If you fix a bug, describe the bug.
We mount many hdfs path to alluxio. When we open the /metrics page ,
find it is very very slow, about 30 mins.
Then debug it , cause of many metrics will invoke isMounteds method one
by one ,and the isMounteds will invoke all mounted path to see it's
capacity twice. But it has no meanings, because we just want to known it
is mounted or not. We can justice it in memory.

Please list the user-facing changes introduced by your change, including
1. change in user-facing APIs no
2. addition or removal of property keys no
3. webui yes, will improve metrics page performance while we have many
mounted path.

pr-link: Alluxio/alluxio#14646
change-id: cid-fe6d4db3b0031cf0b324c8ecdc4d3ccadc3b5ae8