SPDI-31986. Backport Create the test dir earlier in Fuse stressbench cluster mode
We create a directory `fuseIOStressBench` in `mParameters.mLocalPath`
for the stressbench. The whole process should be, first `getattr()`,
find out it doesn't exist, then `create()`.

In cluster mode, job worker 1 may `getattr()` first and find out it
doesn't exist. Before job worker 1 calls `create()`, job worker 2 calls
`create()` first, and thus now the path `/fuseIOStressBench` exists,
which makes worker1 `create()` fail.

With this fix, the node that issues the stressbench command will create
`/fuseIOStressBench` before job service sends the distributed job to job
workers, so that by the time the job worker starts the stressbench, the
path already exists and they won't race on creating this path.

This requires that the node starting the stressbench command also needs
the fuse process just as all other workers, especially if this node is
not a worker node. But we should make sure the correctness first.

pr-link: Alluxio/alluxio#14885
change-id: cid-6f9f88661f7a01dab8ca629a612a7b557f056c21